Initial templates for Ch 2

+++ 1/assets/1.2-finished-code.rs
@@ -5,11 +5,11 @@ use ink_core::storage;
 
 contract! {
     struct Incrementer {
-        value: storage::Value<u32>,
+        value: storage::Value<u64>,
     }
 
     impl Deploy for Incrementer {
-        fn deploy(&mut self, init_value: u32) {
+        fn deploy(&mut self, init_value: u64) {
             self.value.set(init_value)
         }
     }
